fix: export utils from search route
