+++ packages/frontend/src/app/(sticky-header)/search/page.tsx
@@ -6,7 +6,7 @@ import {
   getFilteredSearchResults,
   transferItemsToCards,
   defaultCount,
-} from '@/app/api/search/route'
+} from '@/app/api/search/utils'
 import { EMAIL, ContentType } from '@/app/constants'
 import { LogLevel, log } from '@/app/utils'
